rearranged module structure for easy use
