+++ joglwrapper/__init__.py
@@ -5,3 +5,5 @@ session = requests.Session()
 
 from .reader import Reader
 from .project import Project
+from .member import Member
+from .member_activities import *
